code clean: move puppetResolver to wechaty-puppet, concurrencyExecuter to rx-queue
