+++ package.json
@@ -110,14 +110,16 @@
     "clone-class": "~1.1.1",
     "cockatiel": "^2.0.2",
     "dotenv": "^10.0.0",
+    "file-box": "^1.4.5",
     "fp-ts": "^2.11.5",
     "gerror": "^1.0.14",
     "ix": "^4.5.2",
     "json-rpc-peer": "^0.17.0",
+    "memory-card": "^1.0.3",
     "open-graph": "^0.2.6",
+    "rx-queue": "^1.0.4",
     "state-switch": "~1.6.2",
     "uuid": "^8.3.2",
-    "wechaty-puppet": "~1.11.2",
     "wechaty-puppet-service": "~1.11.1",
     "ws": "^8.2.3"
   },
