--- src/puppet-management/mod.ts
@@ -1,17 +1,0 @@
-import {
-  type PuppetModuleName,
-  PUPPET_DEPENDENCIES,
-  PUPPET_NAME_DEFAULT,
-}                         from './puppet-config.js'
-import {
-  type ResolveOptions,
-  resolvePuppet,
-}                         from './puppet-resolver.js'
-
-export {
-  type PuppetModuleName,
-  type ResolveOptions,
-  PUPPET_DEPENDENCIES,
-  PUPPET_NAME_DEFAULT,
-  resolvePuppet,
-}
