+++ src/wechaty-mixins/misc-mixin.ts
@@ -1,8 +1,9 @@
-import { log }        from 'wechaty-puppet'
+import { log }  from 'wechaty-puppet'
 
 import { VERSION }  from '../config.js'
 
 import type { WechatySkeleton }   from '../wechaty/mod.js'
+
 import type { GErrorMixin }       from './gerror-mixin.js'
 import type { PuppetMixin }       from './puppet-mixin.js'
